+++ android/app/src/test/kotlin/net/mullvad/mullvadvpn/viewmodel/MultihopViewModelTest.kt
@@ -54,7 +54,6 @@ class MultihopViewModelTest {
             WireguardConstraints(
                 isMultihopEnabled = true,
                 entryLocation = Constraint.Any,
-                port = Constraint.Any,
                 ipVersion = Constraint.Any,
             )